Ataymano ataymano/warnings fixes (#1924)
* change semantics of lambda

* namespace hash type fix

* leaf id type fix

* re-fixing hash types

+++ vowpalwabbit/memory_tree.cc
@@ -786,7 +786,7 @@ float return_reward_from_node(memory_tree& b, single_learner& base, uint64_t cn,
 }
 
 void learn_at_leaf_random(
-    memory_tree& b, single_learner& base, const uint32_t& leaf_id, example& ec, const float& weight)
+    memory_tree& b, single_learner& base, const uint64_t& leaf_id, example& ec, const float& weight)
 {
   b.total_num_queries++;
   int32_t ec_id = -1;
